Scanner meta-parsers
Includes three helper functions, Position, Input, and Remaining.
These three functions return the current position of the scanner,
the entire input of the scanner, and the remaining unparsed input
of the scanner respectively.

The helper functions are themselves implemented to satisfy the Parser
interface, specifically they are Parser[int], Parser[string], and
Parser[string] implementations respectively.